Remove file attachment menu
