+++ src/webchat-ui/components/plugins/input/file/DropZone.tsx
@@ -1,14 +1,31 @@
 import React from "react";
 import { styled } from '../../../../style';
+import AttachFile from './attach-file-20px.svg';
 
+const DropZoneContainer = styled.div(({ theme }) => ({
+    position: "absolute",
+    width: "100%",
+    height: theme.unitSize * 12.25,
+    bottom: 0,
+}));
 
-const DropZoneContainer = styled.div(() => ({
-        position: "absolute",
-        boxShadow: "0px 0px 30px 20px rgba(0,0,0,0.1) inset",
-        width: "100%",
-        height: "100%",
-        backgroundColor: 'hsla(0, 0%, 0%, .03)',
-        top: 0  
+const AttachFileIcon = styled(AttachFile)(({ theme }) => ({
+    marginBottom: `-${theme.unitSize / 2}px`,
+}));
+
+const Wrap = styled.div(() => ({
+    alignItems: 'center',
+    justifyContent: 'center',
+    display: 'flex',
+    width: "100%",
+    height: "100%",
+    backgroundColor: "#f7f7f7",
+    boxShadow: "0px 0px 30px 20px rgba(0,0,0,0.1) inset",
+}));
+
+const DragDropTypography = styled.span(() => ({
+    color: 'hsla(0, 0%, 0%, .54)',
+    fontSize: 14,
 }));
 
 interface IDropZoneProps {
@@ -30,7 +47,6 @@ class DropZone extends React.PureComponent<React.HTMLProps<HTMLDivElement> & IDr
 
     handleDragLeave = e => {
         e.preventDefault();
-        e.stopPropagation();
         this.props.setIsDropZoneVisible(false);
     };
 
@@ -46,8 +62,6 @@ class DropZone extends React.PureComponent<React.HTMLProps<HTMLDivElement> & IDr
     }
 
     render() {
-        const { children } = this.props;
-
         return (
             <DropZoneContainer
                 ref={this.dropRef}
@@ -55,7 +69,14 @@ class DropZone extends React.PureComponent<React.HTMLProps<HTMLDivElement> & IDr
                 onDragLeave={e => this.handleDragLeave(e)}
                 onDropCapture={this.handleDrop}
             >
-                {children}
+                <Wrap>
+                    <AttachFileIcon />
+                    <DragDropTypography
+                        className="webchat-input-drag-and-drop-file-text"
+                    >
+                        Drop to attach File
+                    </DragDropTypography>
+                </Wrap>
             </DropZoneContainer>
         );
     }
